Addressing Heitor's feedback
