+++ Makefile
@@ -32,7 +32,7 @@ test:
 	poetry run pytest -m "not perf" --ignore tests/e2e --cov=aws_lambda_powertools --cov-report=xml
 	poetry run pytest --cache-clear tests/performance
 
-nox:
+test-dependencies:
 	poetry run nox --error-on-external-run --reuse-venv=yes --non-interactive
 
 test-pydanticv2:
@@ -50,7 +50,7 @@ coverage-html:
 pre-commit:
 	pre-commit run --show-diff-on-failure
 
-pr: lint lint-docs mypy pre-commit test nox security-baseline complexity-baseline
+pr: lint lint-docs mypy pre-commit test test-dependencies security-baseline complexity-baseline
 
 build: pr
 	poetry build
